updates to perf tests: - updated docs - moved current test runner to default.js - updated run.sh to dump results to results folder - added results folder with placeholder
