+++ test/perf/README.md
@@ -1,9 +1,29 @@
-# Benchmarking Central-Services-Stream
+# Benchmarking
 
-## Docker Compose
+## Pre-requisites
 
-todo.
+Start dependent services as follows:
 
-## Running Tests
+```bash
+docker compose --profile all up -d
+```
 
-todo.
+Stop dependent services as follows:
+
+```bash
+docker compose --profile all down -v
+```
+
+## Running Benchmark Tests
+
+Run tests with standard console output:
+
+```bash
+npm start
+```
+
+Run tests with standard output piped to a log-file captured with todays timestamp - `"%Y%m%d-%Hh%Mm%Ss".log`:
+
+```bash
+sh run.sh
+```
