+++ test/perf/docker.env
@@ -1,1 +1,3 @@
+## Env var configuration for Docker Compose
+KAFKA_REPLICATION_FACTOR=1
 KAFKA_PARTITIONS_NUM=1
