+++ test/perf/run.sh
@@ -6,6 +6,6 @@ now=$(date +"%Y%m%d-%Hh%Mm%Ss")
 echo "Starting test > $now.log"
 
 # Run npm start and pipe the output to a file with the current date and time
-npm start > "$now.log"
+npm start > "./results/$now.log"
 
 echo "Test complete"
